Made UniversePartitioner abstract
