+++ examples/atr/simulation.py
@@ -3,7 +3,7 @@ import openmc
 settings = openmc.Settings()
 settings.batches = 30
 settings.inactive = 10
-settings.particles = 2000
+settings.particles = 40000
 settings.export_to_xml()
 
 openmc.run()
